requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.31
alembic==1.13.2
python-dotenv==1.0.1
structlog==24.2.0
prometheus-client==0.20.0
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-sqlalchemy==0.46b0
opentelemetry-instrumentation-asyncpg==0.46b0
httpx==0.27.0